requirements.txt
attrs==22.2.0
aws-cdk-lib==2.73.0
aws-cdk.asset-awscli-v1==2.2.131
aws-cdk.asset-kubectl-v20==2.1.1
aws-cdk.asset-node-proxy-agent-v5==2.0.107
cattrs==22.2.0
cdk-ecr-deployment==2.5.6
constructs==10.1.304
exceptiongroup==1.1.1
importlib-resources==5.12.0
jsii==1.80.0
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typeguard==2.13.3
typing_extensions==4.5.0
zipp==3.15.0
